Fix interrupting Pie animation (#5879)
## Description

the solution, I'm afraid, is firmly in the userland. We can't do much if
the upstream component changes props on every render.

The `Cell` API is not helping. I have a plan to get rid of it but I will
save it for Recharts@4.

## Related Issue

Closes https://github.com/recharts/recharts/issues/5804

Alternative to https://github.com/recharts/recharts/pull/5839

Bonus: Data changing animation still works

## Screenshots (if appropriate):


https://github.com/user-attachments/assets/b8692425-65ca-4322-b448-6e80f6d79c39