feat: allow minPointSize function to receive null and undefined values (2.x) (#5946)
## Description

A Bar's `minPointSize` attribute can be a function. This PR is aimed at
allowing to pass `null` or `undefined` values to this function (rather
than throwing)

## Related Issue

#5944 

## Motivation and Context

Null or undefined are valid values for datapoints, and are distinct from
`0` (i.e. the absence of data). It's interesting to allow to use those
values even if using a `minPointSize` function.

## How Has This Been Tested?

I tested the fork with a local dataset with null values, and a
minPointSize function returning different values for null non-null
values:

```
const minPointSize = useCallback((value: number | null, _index: number) => isNil(value) ? 0 : 2, []);
```

Other than that, there is little risk of side-effects:
- for existing `minPointSize` functions taking number values, the
function will continue working as is.
- if someone has a codebase sending null values to `minPointSize`,
rechart currently throws.
- minPointSizeCallback is only called at one place in the rechart code,
so there's no other impact
 
## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes