New state types for graphical items (#6148)
## Description

I have inverted the type definition for graphical items. Before, there
was a state type, and each item extended that. Now, each item defines
their types and then the state has union.

The main benefit is that graphical items no longer need to declare
irrelevant properties - i.e. no `stackId` on Pie and no `barSize` on
Line. Also: no more hacks with "select synchronised settings", now it's
just an ID lookup. It is no longer possible for the state, and the props
object, to be out of sync.

Main problem is that this required change in so many places.

Also I converted couple of components to functional variants.

631 errors

## Related Issue

https://github.com/recharts/recharts/issues/5768