fix: Exclude stacked axis domain when not relevant for axis (#6162)
Exclude stacked domain for irrelevant axes. Also added test coverage and
stories.

Co-authored-by: aivarassimkus

## Description

When using an axis of type "number", the stacked domain was always taken
into account, even when the stacking didn't happen on that specific
axis. For example, when we are in a horizontal layout, stacking happens
on the y-axis, so should only be taken into account for those axes. The
opposite holds for vertical layout.

## Related Issue

[6078](https://github.com/recharts/recharts/issues/6078)

## Motivation and Context

In my horizontal area chart, the stacked domain would be included. I'm
using timestamp data (so large numbers like `1753949804`), but the
stacked domain was between 0 and 100. This caused the x-axis domain to
become `[0, 1753949804]`, making the chart display incorrectly.

With this fix, the stacked domain is excluded for the x-axis and only
affects the y-axes.

## How Has This Been Tested?

I've tested this on storybook by adding a story. Also added tests for
area and bar charts (mostly based on the work of @aivarassimkus but also
included one for vertical bar charts).

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] I have added a storybook story or extended an existing story to
show my changes