Add animation tests for Rectangle (#6174)
## Description

Rectangle runs several animations concurrently and our test setup is not
ready for that, I will need to fix that.

Rectangle animations are disabled pretty much everywhere. The only place
I found is in Treemap and in there the animations are ... a disaster. In
this PR I am just documenting/testing the current behaviour, no attempt
on fixing anything.


https://github.com/user-attachments/assets/a4f6510f-34c0-477b-b88e-51e2bbf050d9

## Related Issue

https://github.com/recharts/recharts/issues/5914