Fix Axis blinking when changing props, and YAxis width=auto wrapping (#6262)
## Description

Remove the return null that makes the XAxis disappear in Chrome on weak
hardware (or 4x CPU slowdown)

## More findings

Okay so once I fixed the blink, the text started wrapping with
width=auto. I can fix that by excluding the axis width from tick props
if width=auto (because in that case we want the Text to calculate its
own width, and push it up to the axis). That creates visual diff
elsewhere but that looks more like a bugfix to me so I will do that.

## Related Issue

Fixes https://github.com/recharts/recharts/issues/6257