ResponsiveContainer is built-in to all charts (#6388)
## Description

I never really understood why does this need to be its own element so
let's have these props directly on the main chart shall we.

This also means that now charts without any dimensions at all will
default to 'width=100% height=100%'. Before this PR, chart without
dimensions is just quietly invisible which I think is a disaster. In a
way this is a change of default behaviour - but I prefer to think of it
as a bugfix.

The ResponsiveContainer itself is not as important as before but it
still has some use I suppose - for example in the Bubble story it allows
setting the dimensions from one place. Although that could be a variable
too.