Calling hooks in tests (#6501)
## Description

Tests only.

When putting hooks to `createSelectorTestCase` we used to put them into
`useSelector` which works (hook is a selector without state input).
However such hook is also non memoized and so it's called on each state
recalculation, which gives inaccurate number of re-renders. So let's
call it directly.

Also updated copilot instructions because why tf does the copilot CLI
default to production